Change from BigInteger to Long

+++ src/main/java/com/restful/demo/model/Blog.java
@@ -1,30 +1,33 @@
 package com.restful.demo.model;
 
-import java.math.BigInteger;
-
 public class Blog {
 	
-	private BigInteger id;
+	private Long id;
 	
 	private String title;
 	
 	private String body;
-	
-	public BigInteger getId() {
+
+	public Long getId() {
 		return id;
 	}
-	public void setId(BigInteger id) {
+
+	public void setId(Long id) {
 		this.id = id;
 	}
+
 	public String getTitle() {
 		return title;
 	}
+
 	public void setTitle(String title) {
 		this.title = title;
 	}
+
 	public String getBody() {
 		return body;
 	}
+
 	public void setBody(String body) {
 		this.body = body;
 	}
