+++ src/main/java/com/restful/demo/service/BlogService.java
@@ -1,6 +1,5 @@
 package com.restful.demo.service;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import com.restful.demo.model.Blog;
@@ -9,11 +8,11 @@ public interface BlogService {
 
 	public abstract Collection<Blog> getAll();
 
-	public abstract Blog getId(BigInteger id);
+	public abstract Blog getId(Long id);
 	
 	public abstract Blog create(Blog blog);
 
-	public abstract Blog update(BigInteger id, Blog blog);
+	public abstract Blog update(Long id, Blog blog);
 
-	public abstract boolean delete(BigInteger id);
+	public abstract boolean delete(Long id);
 }
